// app/(authContext)/(platform)/dashboard/kursy/[slug]/strona/types/FixedCardProps.ts
export type FixedCardProps = {
  backgroundColor: string;
  image: File | string | null;
  titleTextColor: string;
  priceTextColor: string;
  discountPriceTextColor: string;
  descriptionTextColor: string;
  featureCheckboxBackgroundColor: string;
  featureCheckboxTextColor: string;
  featureTextColor: string;
  buttonTextColor: string;
  buttonBackgroundColor: string;
};